Player 이동 처리 함수 추가(사유: 플레이어 이동 시 이동 속도에 속도 상수가 제대로 곱해지지 않고 있었음-> 특성 찍어도 변동 없었을 가능성 큼) 스킬 사용 시 순간이동 현상이 원거리 스킬 사용 시에만 발견되는 것 확인 불필요한 패킷 함수 제거(Player Missile -> 발사체 생성하라는 명령을 내리는 패킷, 클라이언트 내부적으로 처리 가능하므로 제거) 자잘한 워닝 제거

+++ LOD_Server/LOD_Server/Code/00.Global/01.Utility/05.CollisionManager/CollisionManager.cpp
@@ -291,7 +291,7 @@ CCollisionObject* CCollisionManager::RequestNearObject(CCollisionObject * pCol, 
 	}
 	return nearObject;
 }
-void CCollisionManager::RequestIncreaseExp(CCollisionObject * pCol, float lengh, TeamType type, float exp)
+void CCollisionManager::RequestIncreaseExp(CCollisionObject * pCol, float lengh, TeamType type, UINT exp)
 {
 	if (m_Winner == TeamType::None){
 		TeamType enemytype;
@@ -326,7 +326,7 @@ void CCollisionManager::RequestIncreaseExp(CCollisionObject * pCol, float lengh,
 						}
 						SC_Msg_Exp_Up p;
 						p.Target_Tag = (short)(*i)->GetTag();
-						p.exp = exp;
+						p.exp = (short)exp;
 						p.size = sizeof(p);
 						p.type = SC_EXP_UP;
 						for (int j = 0; j < MAX_USER; ++j) {
